azdo-api-samples/requirements.txt
requests==2.22.0
python-dotenv==0.10.3
backoff==1.8.0